refactor: extract function and type
